Added New Login Test for Campaign Web Role

+++ src/test/java/SupportingUtilites/GeneralUtilites.java
@@ -2,12 +2,26 @@ package SupportingUtilites;
 import java.security.SecureRandom;
 import java.util.Locale;
 import java.util.Objects;
+import java.util.Properties;
 import java.util.Random;
 import java.io.*;
 
 
 public class GeneralUtilites
 {
+    static Properties prop = new Properties();
+    static String currentPath = System.getProperty("user.dir");
+    static
+    {
+        try
+        {
+            InputStream input = new FileInputStream(currentPath + "/src/test/java/Resources/config.properties");
+            prop.load(input);
+        } catch (Exception e){
+            e.printStackTrace();
+        }
+    }
+
     public static String randomStringGenerator(int intSize)
     {
       return  new RandomString(intSize).nextString();
@@ -42,38 +56,33 @@ public class GeneralUtilites
         }
     }
 
-    public static boolean isProcessRunning(String serviceName) throws Exception {
-
+    public static boolean isProcessRunning(String serviceName) throws Exception
+    {
         Process p = Runtime.getRuntime().exec( "tasklist");
-        BufferedReader reader = new BufferedReader(new InputStreamReader(
-                p.getInputStream()));
+        BufferedReader reader = new BufferedReader
+                (new InputStreamReader(p.getInputStream()));
         String line;
-        while ((line = reader.readLine()) != null) {
-
-            System.out.println(line);
-            if (line.contains(serviceName)) {
+        while ((line = reader.readLine()) != null)
+        {
+           // System.out.println(line);
+            if (line.contains(serviceName))
                 return true;
-            }
         }
-
         return false;
-
     }
 
     public static void KillProcesses()
     {
         try
         {
-            if (isProcessRunning("chromedriver.exe"))
+            if(prop.getProperty("Browser").equals("Chrome"))
             {
-                Runtime.getRuntime().exec("taskkill /F /IM " + "chromedriver.exe");
-            }
+                if (isProcessRunning("chromedriver.exe"))
+                    Runtime.getRuntime().exec("taskkill /F /IM " + "chromedriver.exe");
 
-            if (isProcessRunning("chrome.exe"))
-            {
-                Runtime.getRuntime().exec("taskkill /F /IM " + "chrome.exe");
+                if (isProcessRunning("chrome.exe"))
+                    Runtime.getRuntime().exec("taskkill /F /IM " + "chrome.exe");
             }
-
         }
         catch (Exception e)
         {
